probe/sniff: move to experimental
Resolves #466
